requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
whitenoise==6.7.0
sqlparse==0.5.1
asgiref==3.8.1
PyJWT==2.9.0
flake8==7.0.0
pytest==8.1.1
pytest-django==4.8.0    